renamed views directory
