+++ assets/javascripts/components/message-input/template.vue
@@ -2,4 +2,4 @@
     <textarea ref="sayWidget" name="say_text" id="say_text" placeholder="Let's talk!" v-bind:rows="inputLines" v-bind:disabled="disabled" v-model="text" v-on:keypress="maybeSay($event)"></textarea>
 </template>
 
-<script src="./message-input.js"></script>
+<script src="./script.js"></script>
