+++ assets/javascripts/components/messages/script.js
