+++ assets/javascripts/components/rooms/script.js
