+++ assets/javascripts/components/rooms/style.css
